Add new TextLabel storybook doc
